==> 01_exploración_de_datos.py <==
# -*- coding: utf-8 -*-
"""01 - Exploración de Datos.ipynb

Automatically generated by Colab.

Original file is located at
    https://colab.research.google.com/drive/1wyg-7JHfg_dU5j2_s6gSttFmoenYP5TR

# Exploración de Datos

Este notebook carga y explora los datos del dataset de resultados de futbol internacional entre los años 1872 y 2017.

Los archivos son:

1. result.csv
2. goalscorers.csv
3. former_names.csv
4. shootouts.csv

Los datos se cargan desde un archivo .zip local subidos manualmente
"""

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from google.colab import files

"""Cargar el archivo .zip del Dataset"""

uploaded = files.upload()

"""Se descomprime el archivo"""

!unzip archive.zip

"""Cargar los Datasets"""

df_results = pd.read_csv('results.csv')
df_goalscorers = pd.read_csv('goalscorers.csv')
df_former_names = pd.read_csv('former_names.csv')
df_shootouts = pd.read_csv('shootouts.csv')

"""Mostrar la dimensiones de cada set"""

print("Resultados: ", df_results.shape)
print("Goleadores: ", df_goalscorers.shape)
print("Nombres Antiguos: ", df_former_names.shape)
print("Penales: ", df_shootouts.shape)

"""Vista inicial de cada Dataset"""

df_results.head()

df_goalscorers.head()

df_former_names.head()

df_shootouts.head()

"""Información general de cada Dataset"""

df_results.info()

df_goalscorers.info()

df_former_names.info()

df_shootouts.info()

"""Resumen estadistico"""

df_results.describe(include='all')

df_goalscorers.describe(include='all')

df_former_names.describe(include='all')

df_shootouts.describe(include='all')

"""Búsqueda de valores nulos"""

print("Nulos en results.csv:")
display(df_results.isnull().sum())

print("\nNulos en goalscorers.csv:")
display(df_goalscorers.isnull().sum())

print("\nNulos en former_names.csv:")
display(df_former_names.isnull().sum())

print("\nNulos en shootouts.csv:")
display(df_shootouts.isnull().sum())

"""Búsqueda de valores duplicados"""

print("Duplicados en results: ", df_results.duplicated().sum())
print("Duplicados en goalscorers: ", df_goalscorers.duplicated().sum())
print("Duplicados en former_names: ", df_former_names.duplicated().sum())
print("Duplicados en shootouts: ", df_shootouts.duplicated().sum())

"""Gráfica de distribución de goles"""

plt.figure(figsize=(8,5))
sns.histplot(df_results['home_score'], kde = True)
plt.title("Distribución de goles del equipo local")
plt.show()

plt.figure(figsize=(8,5))
sns.histplot(df_results['away_score'], kde = True)
plt.title("Distribución de goles del equipo visitante")
plt.show()

"""Número de partidos por año"""

df_results['year'] = pd.to_datetime(df_results['date']).dt.year

plt.figure(figsize=(12,6))
df_results['year'].value_counts().sort_index().plot()
plt.title("Número de partidos por año")
plt.xlabel("Año")
plt.ylabel("Cantidad de partidos")
plt.show
